pyproject.toml
[project]
name = "enph353-training"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "modal>=1.2.2",
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
    "ultralytics>=8.3.228",
    "wandb>=0.23.0",
    "pyyaml>=6.0.2",
]
